src/python_wrapper/requirements.txt
PyYAML==6.0
a2wsgi==1.7.0
httpx==0.24.0
